Create calculate-gw-subnet
A virtual network (VNet) in Azure infrastructure services that is connected to other networks must have a gateway subnet. The best practices for defining this subnet are the following:

The prefix length of the gateway subnet can have a maximum prefix length of 29 (for example, 10.119.255.248/29), but the current recommendation is that you use a prefix length of 27 (for example, 10.119.255.224/27).
When defining the address space of the gateway subnet, use the very last part of the VNet address space.
For the second recommendation, you can determine the address space of the gateway subnet by setting the bits used for the gateway subnet to 0 and the remaining variable bits in the VNet address space to 1. To quickly calculate the gateway subnet address space without having to convert to binary and back to decimal, you can use GWSubnetCalc.cs, a console application written in C#, or the PowerShell command block above.

This C# or Python console application and PowerShell command block collects five integers—the values of w.x.y.z/n for the VNet address prefix and the gateway subnet prefix length—and calculates the gateway subnet address space.

Please note that there is limited error checking of the input values, so improper values will most likely produce improper results.